src/requirements.txt
ConfigParser
requests
kafka-python
python-daemon
